Tidy up a bit.
